requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
pydantic==2.8.2
google-generativeai>=0.8.3
sqlalchemy==2.0.31
asyncpg==0.29.0
